refactor: move method index to own file, add methodUnknown zero value
